Complete project polish and testing infrastructure
- Organize test suite into tests/ directory structure
- Add comprehensive stress and performance testing
- Remove legacy test files and cleanup project structure
- Update documentation to be developer-friendly
- Polish Makefile with proper test targets
- Remove emojis from all output for professional appearance
- Add Docker test runner for Linux compatibility

+++ src/core/storage.c
@@ -176,7 +176,7 @@ int storage_put(const char* key, const char* value, size_t value_size) {
     
     // Calculate blocks needed based on data area size, not full struct size
     int blocks_needed = (value_size + sizeof(((struct data_block*)0)->data) - 1) / sizeof(((struct data_block*)0)->data);
-    if (blocks_needed > meta.free_blocks) {
+    if ((uint32_t)blocks_needed > meta.free_blocks) {
         return -1;  // Not enough space
     }
